+++ src/solarGenerator/solarSystemGen.cpp
@@ -94,7 +94,7 @@ void genBigSizePlanets(ostringstream* solar){
         switch (i) {
             case 5:
                 for(int j {0}; j < 3; j++){
-                    float prp { (1.0f/plProp[i]) };
+                    float prp { static_cast<float>(1.0f/plProp[i]) };
                     *solar << "         <group>" << endl;
                     *solar << "             <scale x=\"" << prp << "\" y=\"" << prp << "\" z=\"" << prp << "\" />" << endl;
                     *solar << "             <rotate angle=\"" << angleMoons << "\" axisX=\"0\" axisY=\"1\" axisZ=\"0\" />" << endl;
@@ -151,7 +151,7 @@ void genTrajectories(ostringstream* solar){
     dumpFile(file, pF);
 
     for(int i { 0 }; i < 9; i++){
-        float prop { plD[i]/plD[2] };
+        float prop { static_cast<float>(plD[i]/plD[2]) };
         *solar << "     <group>" << endl;
         *solar << "         <scale x=\"" << prop << "\" y=\"" << prop << "\" z=\"" << prop << "\" />" << endl;
         *solar << "         <models>" << endl;
